requirements.txt
rich>=13.0.0
questionary>=2.0.0
pyyaml>=6.0
pydantic>=2.0.0
mysql-connector-python>=8.0.0
psutil>=5.9.0
python-nmap>=0.7.1
